// src/app/api/chat/constants.ts
// eslint-disable-next-line import/prefer-default-export
export enum ThreadType {
  group = 'group',
  direct = 'direct',
}

export enum ChatMessageType {
  text = 'text',
  html = 'html',
  topicUpdated = 'topicUpdated',
  participantAdded = 'participantAdded',
  participantRemoved = 'participantRemoved',
}